split src/utils.ts into separate common/*.ts categories
> Aside: Make a util directory and keep different utilities in
> different files. A single util file will always grow until it is
> too big and yet too hard to split apart. Using a single util file
> is unhygienic.

https://programmingisterrible.com/post/139222674273
