+++ src/common/auth.ts
@@ -1,10 +1,7 @@
 import * as jwt from 'jsonwebtoken';
-import { Prisma } from './generated/prisma';
 
-export interface Context {
-  db: Prisma;
-  request: any;
-}
+import Context from './context';
+import { AuthError } from './error';
 
 export function getUserId(ctx: Context) {
   const Authorization = ctx.request.get('Authorization');
@@ -17,10 +14,4 @@ export function getUserId(ctx: Context) {
   }
 
   throw new AuthError();
-}
-
-export class AuthError extends Error {
-  constructor() {
-    super('Not authorized');
-  }
 }
