+++ src/common/context.ts
@@ -1,0 +1,6 @@
+import { Prisma } from '../generated/prisma';
+
+export default interface Context {
+  db: Prisma;
+  request: any;
+}
