+++ src/resolvers/AuthPayload.ts
@@ -1,4 +1,4 @@
-import { Context } from '../utils';
+import Context from '../common/context';
 
 export const AuthPayload = {
   user: async ({ user: { id } }, args, ctx: Context, info) => {
